pyproject.toml
[project]
name = "utility-bills"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "pydantic==2.11.7",
    "python-telegram-bot==22.2",
]
